Refactor tests to require explicit log in
The "log in" action should be explicitly triggered from the tests in
order to help confirm consistent application behavior over time.

In addition to being more explicit, this approach creates a logical
place to stub the initial request for project phase data. This request
is dynamic (based on the local system time), so it must be stubbed out
to facilitate deterministic test execution with pre-recorded
request/response pairs.