Remove authorization logic
This client authorization has been ported to the API server, so the
application code can be simplified to rely on that service.

Update the stubbing mechanism to continue to support offline development
and testing.

Fully parameterize authentication providers to simplify the process of
adding/removing providers in the future.

Update the fixture data to use a new access token (because the access
token is now specified via a query string parameter and not a
'Set-cookie' header, whitespace should not be used).