Avoid invalid state when updating multiple models
A single "save" operation on a Utilizations collection may include the
expansion and contraction of adjacent models. Because each model is
saved in isolation, issuing all "update" requests in parallel may result
in an invalid state on the server (one where two different utilization
models include the same day).

Introduce a private `safeUpdate` function to intelligently (if
inefficiently) batch requests to avoid this problem.