alter disabled selection criteria
Previously, the UI disabled selects for position, project and
phase if the selected utilization type was id #1 (consulting).

Utilization types have a field, `project_required`, which is a
more accurate heuristic for this. This commit implements support
for that field.

I've also removed the tests for the utilization_type model as it
no longer has derived properties to be tested. I didn't see any
tests for the UI behavior, and have not attempted to add any.

closes gh-59