Fix regression in drag-and-drop operation
Commit 0657913e30f8edf26d66571625534b58a77c06b6 introduced a regression
in the utilization drag-and-drop feature. Specifically, it broke
everything. Fix the bug and add a UI test to prevent future regressions.

Some combination of Firefox, Selenium, and Leadfoot prohibits automated
testing of drag-and-drop interactions. Introduce the necessary logic to
assert the correctness of this change, but disable it until the
underlying testing problem can be identified and resolved.