Render utilization phase name in utilization tile
The `Utilization` model does not define a `phase` property. This means
that the name `phase` refers to the project phase of the current review
(and *not* the project phase of the current utilization) From the
`utilization-row-day` component (even within a context-setting `#with`
directive).

Update the "utilization tile" view to display the utilization phase
rather than the review phase.