[Fix] Correct input form for phaseless projects
The "project phase" input options are serialized from multiple
responses. This means that the strict equality operator cannot be
reliably used to compare the likeness of a "phase" input option and a
utilization model's "phase" attribute.

Phase model ID values are primitive JavaScript numbers that describe the
identity of the phase. This makes them suitable for comparison across
objects that, despite being distinct in memory, describe the same
"phase" model.

Update the "edit utilization" form to initialize its values according to
model IDs, not model objects.