[Fix] Serve application with correct headers
By neglecting to specify the `Content-Type` header, the client (or the
proxy between the server and the client) must infer the MIME type of the
optimized application file.

In production, an SSL terminator serves as a proxy between this server
and the outside world. As part of its job enforcing consistent security
settings, it (incorrectly) infers the `Content-Type` of this file and
(correctly) sets the `X-Content-Type-Options` header to "nosniff". This
combination of changes causes Internet Explorer and Google Chrome to
refuse to execute the JavaScript code.

Explicitly set the `Content-Type` header to the correct value to avoid
this issue.